requirements.txt
fastapi
uvicorn
sqlmodel
sqlalchemy
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pbkdf2
python-dotenv
